\name{CheckDigits}
\title{Create encoded serial numbers for European Health Examination Surveys (EHES)}
\alias{CheckDigits}
\alias{Check}
\alias{ControlDigit}
\alias{CreateDigits}
\alias{Perm.fn}
\alias{Add.fn}
\alias{digits}
\alias{undigits}
\alias{Check.number}
\alias{Check.length}
\alias{ListCheck}
\alias{Add.control}
\alias{Create.number}
\alias{Create.control}
\alias{dataPreview}
\description{These functions allow the user to create and check encoded serial numbers. The encoding uses the dihedral group of order 10 for addtion method as described by Gumm (1986). }
\references{See the EHES manual published on the EHES website: http://www.ehes.info
Gumm (1986) Encoding of numbers to detect typing errors. International Journal of Applied Engineering Education 2(1)}
\author{Statistics Norway\cr
Maintainer: Susie.Cooper@ssb.no & Johan.Heldal@ssb.no}
\keyword{ survey }