\name{digi}
\alias{digi}
\docType{data}
\title{Internal dataset used to create encoded serial numbers}
\description{
	This dataset is in three parts, all of which are used to 
	create and check encoded serial numbers. The encoding uses 
	the dihedral group of order 10 for addtion method as 
	described by Gumm (1986). }
\usage{digi}
\format{Two datasets and a vector, strung together as a list.}
\source{Gumm (1986) Encoding of numbers to detect typing errors. 
	International Journal of Applied Engineering Education 2(1)}
\references{See the EHES manual published on the EHES website: http://www.ehes.info}
\keyword{ datasets }