\name{EHESsampling}
\title{Functions to assist with sampling for European Health Examination Surveys (EHES)}
\alias{SampleSizes}
\alias{stage1Sample.fn}
\alias{stage2Sample.fn}
\alias{mk.calc}
\alias{special.round}
\description{This provides tools to assist with European Health Examination Surveys, including calculating sample sizes and sampling with a 2-stage design.
A detailed user-friendly manual has been created. Please contact one of the maintainers below for a copy of this.}
\references{see the EHES manual published on the EHES website:
http://www.ehes.info}
\author{Statistics Norway\cr
Maintainer: susie.jentoft@ssb.no & johan.heldal@ssb.no}
\keyword{ survey }
