\name{RcmdrPlugin.EHESsampling - internal}
\title{Internal functions for menu system of RcmdrPlugin.EHESsampling}
\alias{OpenSampling1}
\alias{OpenSampling2}
\alias{OpenSampleSizes}
\alias{OpenCheck}
\alias{OpenCreateDigits}
\alias{OpenControlDigit}
\alias{Sample.notebook}
\alias{Stage1.notebook}
\alias{Stage1.notebook2}
\alias{Stage2.notebook}
\alias{Stage2.notebook2}
\alias{info.bn}
\alias{DataPreview}
\alias{disable.fn}
\alias{disable.fn2}
\alias{disable.fn3}
\alias{setstate}
\alias{setstate2}
\alias{OKCancelHelp2}
\alias{PSU.options}
\alias{Sample.datasets}
\alias{Stage2onHelp}
\alias{Stage2.options}
\alias{UPMEpiktildedrompik1}
\alias{UPmaxentopy1}
\alias{as_int1}
\alias{check.fn}
\alias{check.fn2}
\alias{costvar.options}
\alias{mergeDataSets}
\alias{tkblank}
\alias{radioButtons2}
\alias{strata.options}
\alias{variableListBox2}
\alias{ALL.options}
\alias{PSUonHelp}
\alias{Stage1onHelp}
\alias{UPMEpiktildefrompik1}
\alias{UPmaxentropy1}
\alias{costonHelp}
\alias{dataonHelp}
\alias{reportonHelp}
\alias{strataonHelp}


\description{This provides tools to assist with European Health Examination Surveys, including calculating sample sizes and sampling with a 2-stage design.
A detailed user-friendly manual has been created. Please contact one of the maintainers below for a copy of this.}
\references{see the EHES manual published on the EHES website: http://www.ehes.info}
\author{Statistics Norway\cr
Maintainer: susie.jentoft@ssb.no & johan.heldal@ssb.no}
\keyword{ survey }
