\name{RcmdrPlugin.EHESsampling-package}
\alias{RcmdrPlugin.EHESsampling-package}
\docType{package}
\title{Help with EHES sampling}
\description{This package provides a range of tools to assist with the sampling process for European Health Examination Surveys (EHES). Tools include programs to calculate sample sizes, selecting a 2-stage sample, importing a SAS dataset and creating encoded serial numbers for sample units.}
\details{\tabular{ll}{
Package: \tab RcmdrPlugin.EHESsampling\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2010-05-18\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr}
}
\author{Statistics Norway\cr
Maintainer: susie.jentoft@ssb.no & johan.heldal@ssb.no}
\references{For an indepth manual please email the authors\cr
The general EHES manual is located publically on the EHES website: www.ehes.info}
\keyword{ package }
