\name{RcmdrPlugin.EHESsampling - internal}
\title{Internal functions for menu system of RcmdrPlugin.EHESsampling}
\alias{OpenSampling1}
\alias{OpenSampling2}
\alias{OpenWeights}
\alias{OpenSampleSizes}
\alias{OpenCheck}
\alias{OpenCreateDigits}
\alias{OpenControlDigit}
\alias{Sample.notebook}
\alias{Stage1.notebook}
\alias{Stage2.notebook}
\alias{info.bn}
\alias{DataPreview}
\alias{disable.fn}
\alias{setstate}
\description{This provides tools to assist with European Health Examination Surveys, including calculating sample sizes and sampling with a 2-stage design.
A detailed user-friendly manual has been created. Please contact one of the maintainers below for a copy of this.}
\references{see the EHES manual published on the EHES website: http://www.ehes.info}
\author{Statistics Norway\cr
Maintainer: susie.cooper@ssb.no & johan.heldal@ssb.no}
\keyword{ survey }
